+++ src/LayoutManager/resolvers/ClipPathLayoutResolver.ts
@@ -1,32 +1,33 @@
 import { Point } from '../../Point';
 import type { Group } from '../../shapes/Group';
 import type { FabricObject } from '../../shapes/Object/FabricObject';
-import { invertTransform } from '../../util';
-import type { LayoutContext, LayoutStrategyResult } from '../types';
+import { sendPointToPlane } from '../../util';
+import type { LayoutResolverResult, StrictLayoutContext } from '../types';
 import { LayoutResolver } from './LayoutResolver';
 
 export class ClipPathLayoutResolver extends LayoutResolver {
   calcLayoutResult(
     target: Group,
     objects: FabricObject[],
-    context: LayoutContext
-  ): LayoutStrategyResult | undefined {
+    context: StrictLayoutContext
+  ): LayoutResolverResult | undefined {
     const clipPath = target.clipPath;
     if (!clipPath) {
       return;
     }
+    const isLayoutChange =
+      context.prevResolver && context.resolver !== context.prevResolver;
     const clipPathSizeAfter = clipPath._getTransformedDimensions();
     if (
       clipPath.absolutePositioned &&
-      (context.type === 'initialization' || context.type === 'layout_change')
+      (context.type === 'initialization' || isLayoutChange)
     ) {
       //  we want the center point to exist in group's containing plane
-      let clipPathCenter = clipPath.getCenterPoint();
-      if (target.group) {
-        //  send point from canvas plane to group's containing plane
-        const inv = invertTransform(target.group.calcTransformMatrix());
-        clipPathCenter = clipPathCenter.transform(inv);
-      }
+      const clipPathCenter = sendPointToPlane(
+        clipPath.getCenterPoint(),
+        undefined,
+        target.group?.calcTransformMatrix()
+      );
       return {
         centerX: clipPathCenter.x,
         centerY: clipPathCenter.y,
@@ -34,19 +35,15 @@ export class ClipPathLayoutResolver extends LayoutResolver {
         height: clipPathSizeAfter.y,
       };
     } else if (!clipPath.absolutePositioned) {
-      let center;
       const clipPathRelativeCenter = clipPath.getRelativeCenterPoint(),
         //  we want the center point to exist in group's containing plane, so we send it upwards
         clipPathCenter = clipPathRelativeCenter.transform(
           target.calcOwnMatrix(),
           true
         );
-      if (
-        context.type === 'initialization' ||
-        context.type === 'layout_change'
-      ) {
+      if (context.type === 'initialization' || isLayoutChange) {
         const bbox = this.calcBoundingBox(target, objects, context) || {};
-        center = new Point(bbox.centerX || 0, bbox.centerY || 0);
+        const center = new Point(bbox.centerX || 0, bbox.centerY || 0);
         return {
           centerX: center.x + clipPathCenter.x,
           centerY: center.y + clipPathCenter.y,
@@ -56,7 +53,7 @@ export class ClipPathLayoutResolver extends LayoutResolver {
           height: clipPath.height,
         };
       } else {
-        center = target.getRelativeCenterPoint();
+        const center = target.getRelativeCenterPoint();
         return {
           centerX: center.x + clipPathCenter.x,
           centerY: center.y + clipPathCenter.y,
