+++ src/LayoutManager/resolvers/FitContentLayoutResolver.ts
@@ -1,6 +1,6 @@
 import type { Group } from '../../shapes/Group';
 import type { FabricObject } from '../../shapes/Object/FabricObject';
-import type { LayoutContext, LayoutStrategyResult } from '../types';
+import type { LayoutResolverResult, StrictLayoutContext } from '../types';
 import { LayoutResolver } from './LayoutResolver';
 
 export class FitContentLayoutResolver extends LayoutResolver {
@@ -9,8 +9,8 @@ export class FitContentLayoutResolver extends LayoutResolver {
   calcLayoutResult(
     target: Group,
     objects: FabricObject[],
-    context: LayoutContext
-  ): LayoutStrategyResult | undefined {
+    context: StrictLayoutContext
+  ): LayoutResolverResult | undefined {
     const targetsToMeasure =
       this.lazy &&
       context.type === 'added' &&
